requirements.txt
Django==4.2.4
psycopg2-binary==2.9.7
Pillow==10.0.0
django-jazzmin==2.6.0
django-shortuuidfield==0.1.3
flake8==6.1.0
